% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRConstruction.R
\name{sr}
\alias{sr}
\title{Create SR given one scanpath}
\usage{
sr(data, nAOI, a = 0, g = 0, converged = FALSE, mat = TRUE)
}
\arguments{
\item{data}{Matrix or dataframe with two columns (first column is the sender (i), second receiver (j))}

\item{nAOI}{positive integer indicating the number of AOIs}

\item{a}{alpha. Numeric between 0 and 1}

\item{g}{gamma. Numeric between 0 and 1}

\item{converged}{Logical. Should the SR matrix be constructed statically based on first order transition matrix?}

\item{mat}{Logical. Should the output be matrix or vector? Used for internal purposes.}
}
\value{
matrix or vector of the Successor Representation values. The vector is columnwise unfold of a matrix.
}
\description{
This function takes two vectors of states (i and j) and converts it into Successor Representation of the transition sequence.
}
