% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRConstruction.R
\name{srs}
\alias{srs}
\title{Create a Successor Representations}
\usage{
srs(data, nAOI, a, g, converged = FALSE, averageItems = TRUE,
  normalize = FALSE)
}
\arguments{
\item{data}{matrix or dataframe of 4 columns. 1 Id of participant, 2 item, 3 sender and 4 receiver}

\item{nAOI}{number of Areas of interest (positive integer)}

\item{a}{alpha. Numeric between 0 and 1}

\item{g}{gamma. Numeric between 0 and 1}

\item{converged}{Logical. Should the SR matrix be constructed statically based on first order transition matrix?}

\item{averageItems}{Logical. Should the matrices be averaged across items?}

\item{normalize}{Logical. Should the matrix cells be normalized?}
}
\description{
Create a Successor Representations
}
